+++ source/common/quic/BUILD
@@ -673,6 +673,18 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "envoy_deterministic_connection_id_generator_lib",
+    srcs = ["envoy_deterministic_connection_id_generator.cc"],
+    hdrs = ["envoy_deterministic_connection_id_generator.h"],
+    tags = ["nofips"],
+    deps = [
+        ":envoy_quic_connection_id_generator_factory_interface",
+        ":envoy_quic_utils_lib",
+        "@com_github_google_quiche//:quic_core_deterministic_connection_id_generator_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "envoy_quic_server_preferred_address_config_factory_interface",
     hdrs = ["envoy_quic_server_preferred_address_config_factory.h"],
     tags = ["nofips"],
