+++ source/extensions/quic/connection_id_generator/deterministic/BUILD
@@ -16,25 +16,13 @@ licenses(["notice"])  # Apache 2
 envoy_extension_package()
 
 envoy_cc_library(
-    name = "envoy_deterministic_connection_id_generator_lib",
-    srcs = ["envoy_deterministic_connection_id_generator.cc"],
-    hdrs = ["envoy_deterministic_connection_id_generator.h"],
-    tags = ["nofips"],
-    deps = [
-        "//source/common/quic:envoy_quic_connection_id_generator_factory_interface",
-        "//source/common/quic:envoy_quic_utils_lib",
-        "@com_github_google_quiche//:quic_core_deterministic_connection_id_generator_lib",
-    ],
-)
-
-envoy_cc_library(
     name = "envoy_deterministic_connection_id_generator_config_lib",
     srcs = ["envoy_deterministic_connection_id_generator_config.cc"],
     hdrs = ["envoy_deterministic_connection_id_generator_config.h"],
     tags = ["nofips"],
     deps = [
-        ":envoy_deterministic_connection_id_generator_lib",
         "//envoy/registry",
+        "//source/common/quic:envoy_deterministic_connection_id_generator_lib",
         "//source/common/quic:envoy_quic_connection_id_generator_factory_interface",
         "@envoy_api//envoy/extensions/quic/connection_id_generator/v3:pkg_cc_proto",
     ],
